#3896 - Persisted queries that are exported throw an error when trying to import
git-svn-id: https://subversion.assembla.com/svn/mdss@9105 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: bfb5efd6a3c26c61c7577d2e717d71fb9861ec84 [formerly c48acacddd593eab40e4d0f9755a4680aa8089d4]
Former-commit-id: a087df51c1cc14c1636654e413e6f5208fb41278

+++ trunk/src/server/dss/vector/solutions/query/MdTableBuilder.java
@@ -655,22 +655,25 @@ public class MdTableBuilder implements Reloadable
         {
           BrowserField field = BrowserField.getBrowserField(mdAttributeReference);
 
-          OIterator<? extends BrowserRoot> it = field.getAllroot();
-          try
+          if (field != null)
           {
-            List<? extends BrowserRoot> roots = it.getAll();
-
-            for (BrowserRoot root : roots)
+            OIterator<? extends BrowserRoot> it = field.getAllroot();
+            try
             {
-              root.delete();
-            }
-          }
-          finally
-          {
-            it.close();
-          }
+              List<? extends BrowserRoot> roots = it.getAll();
 
-          field.delete();
+              for (BrowserRoot root : roots)
+              {
+                root.delete();
+              }
+            }
+            finally
+            {
+              it.close();
+            }
+
+            field.delete();
+          }
         }
       }
     }
